Делит код на отдельные ES2015-модули
